localized classification of resolver errors
